Split focal to focal, focal-state and focal-utils
