+++ package.json
@@ -2,10 +2,10 @@
   "private": true,
   "name": "focal",
   "scripts": {
-    "build": "yarn workspace @grammarly/focal build",
-    "package": "yarn workspace @grammarly/focal pack",
-    "release": "yarn workspace @grammarly/focal release",
-    "test": "yarn workspace @grammarly/focal test && yarn workspace focal-todomvc build && yarn workspace focal-examples build && yarn workspace focal-manual-tests build",
+    "build": "yarn workspace @grammarly/focal-state build && yarn workspace @grammarly/focal build ",
+    "package": "yarn workspace @grammarly/focal-state pack && yarn workspace @grammarly/focal pack",
+    "release": "yarn workspace @grammarly/focal-state release && yarn workspace @grammarly/focal release",
+    "test": "yarn workspace @grammarly/focal-state test && yarn workspace @grammarly/focal test && yarn workspace @grammarly/focal-utils test && yarn workspace focal-todomvc build && yarn workspace focal-examples build && yarn workspace focal-manual-tests build",
     "postinstall": "yarn build"
   },
   "devDependencies": {
@@ -15,6 +15,8 @@
   },
   "workspaces": [
     "packages/focal",
+    "packages/focal-state",
+    "packages/focal-utils",
     "packages/test",
     "packages/examples/*"
   ],
