+++ packages/focal-state/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@grammarly/focal",
+  "name": "@grammarly/focal-state",
   "version": "0.7.0",
   "description": "FRP UI with React, observables, immutable data and lenses",
   "main": "dist/src/index.js",
@@ -79,14 +79,9 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@grammarly/focal-state": "0.7.0",
     "@grammarly/focal-utils": "0.1.0",
     "@grammarly/tslint-config": "0.5.1",
     "@types/node": "^6.0.34",
-    "@types/react": "16.0.33",
-    "@types/react-dom": "16.0.3",
-    "react": "16.2.0",
-    "react-dom": "16.2.0",
     "tslint": "5.2.0",
     "typescript": "2.8.4",
     "rxjs": "5.5.6",
@@ -96,11 +91,6 @@
     "watch": "^0.18.0"
   },
   "peerDependencies": {
-    "@grammarly/focal-state": ">= 0.7.0 < 0.8.0",
-    "@types/react": ">= 16.0.33 < 17.0.0-0",
-    "@types/react-dom": ">= 16.0.3 < 17.0.0-0",
-    "react": ">= 16.2.0 < 17.0.0-0",
-    "react-dom": ">= 16.2.0 < 17.0.0-0",
     "rxjs": ">= 5.4.2 < 6.0.0-0"
   }
 }
