+++ packages/focal-state/src/lens/base.ts
