+++ packages/focal-state/test/index.ts
@@ -1,0 +1,2 @@
+import './test_lens'
+import './test_atom'
