+++ packages/focal-utils/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@grammarly/focal-utils",
+  "private": true,
+  "version": "0.1.0",
+  "description": "FRP UI with React, observables, immutable data and lenses",
+  "main": "dist/src/index.js",
+  "typings": "dist/src/index.d.ts",
+  "files": [
+    "dist/src/"
+  ],
+  "scripts": {
+    "docs": "rm -rf ./docs && typedoc --out docs --theme minimal --ignoreCompilerErrors --tsconfig tsconfig.json",
+    "clean": "rm -rf ./dist",
+    "build": "npm run clean && tsc && npm run lint",
+    "test": "npm run build",
+    "test:watch": "watch 'npm test' src test",
+    "lint": "tslint --type-check --project tsconfig.json -c tslint.json"
+  },
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/grammarly/focal.git"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@grammarly/tslint-config": "0.5.1",
+    "tslint": "5.2.0",
+    "typescript": "2.8.4",
+    "@types/tape": "^4.2.27",
+    "tap-spec": "^4.1.1",
+    "tape": "^4.5.1",
+    "watch": "^0.18.0"
+  }
+}
