+++ packages/focal-utils/tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@grammarly/tslint-config",
+  "rules": {
+    "strict-boolean-expressions": [false]
+  }
+}
