+++ packages/focal/src/index.ts
@@ -1,6 +1,4 @@
-export { Lens, Prism, Optic } from './lens/index'
-export { Atom, ReadOnlyAtom } from './atom/index'
-export { Option } from './utils'
+export { Lens, Prism, Optic, Atom, ReadOnlyAtom, Option } from '@grammarly/focal-state'
 
 export {
   bind,
