+++ packages/focal/test/index.ts
@@ -1,3 +1,1 @@
-import './test_lens'
-import './test_atom'
 import './test_react'
